# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /oauth/google/login:
    get:
      description: Redirects to Google OAuth authorization endpoint
      operationId: OAuthController_googleLogin
      parameters:
        - name: redirect
          required: true
          in: query
          schema:
            type: string
        - name: frontend
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Initiate Google OAuth login flow
      tags:
        - oauth
...
  /oauth/google/callback:
    get:
      description: Handles the OAuth callback from Google
      operationId: OAuthController_googleCallback
      parameters:
        - name: code
          required: true
          in: query
          schema:
            type: string
        - name: state
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Google OAuth callback endpoint
      tags:
        - oauth
  /oauth/microsoft/login:
    get:
      description: Redirects to Microsoft OAuth authorization endpoint
      operationId: OAuthController_microsoftLogin
      parameters:
        - name: redirect
          required: true
          in: query
          schema:
            type: string
        - name: frontend
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Initiate Microsoft OAuth login flow
      tags:
        - oauth
...
  /oauth/microsoft/callback:
    get:
      description: Handles the OAuth callback from Microsoft
      operationId: OAuthController_microsoftCallback
      parameters:
        - name: code
          required: true
          in: query
          schema:
            type: string
        - name: state
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      summary: Microsoft OAuth callback endpoint
      tags:
        - oauth
  /oauth/google/unlink:
    post:
      description: >-
        Removes Google identity from current user account. Requires at least one
        other identity to remain.
      operationId: OAuthController_unlinkGoogle
      parameters: []
      responses:
        '204':
          description: Google account unlinked successfully
        '400':
          description: Cannot unlink - user must have at least one identity
        '404':
          description: Google account not linked
      security:
        - bearer: []
      summary: Unlink Google account from current user
      tags:
        - oauth
  /oauth/microsoft/unlink:
    post:
      description: >-
        Removes Microsoft identity from current user account. Requires at least
        one other identity to remain.
      operationId: OAuthController_unlinkMicrosoft
      parameters: []
      responses:
        '204':
          description: Microsoft account unlinked successfully
        '400':
          description: Cannot unlink - user must have at least one identity
        '404':
          description: Microsoft account not linked
      security:
        - bearer: []
      summary: Unlink Microsoft account from current user
      tags:
        - oauth
...
info:
  title: Todo App API
  description: >-
    REST API for Todo Application with authentication and role-based access
    control
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...